Fix incorrect heap size value for boxed slice
Previously the `GetSize` impl for a boxed slice didn't have a bound on `T`, so
`get_size` was being called on `&T` instead of `T`. This meant that the size of
each element was always being calculated to the size of a reference. This change
fixes that and adds a test.

Co-Authored-By: Troy Hinckley <t.macman@gmail.com>